jm/GH-61-faucet-api (#1)
* First successful build

* Successful send action

* PR comments
