+++ .env.example
@@ -1,1 +1,4 @@
-PRIVATE_KEY=<PRIVATE_KEY>
+PRIVATE_KEY=<PRIVATE-KEY>
+PUBLIC_ACTOR=feengfaucet1
+PUBLIC_PERMISSION=active
+PUBLIC_BROADCAST=true
