+++ src/lib/constants.ts
@@ -1,9 +1,10 @@
-export const ACTOR = process.env.ACTOR ?? 'eosio.faucet';
-export const ACCOUNT = process.env.ACCOUNT ?? ACTOR;
-export const EVM = 'eosio.evm';
-export const FAUCET = 'eosio.faucet';
-export const PERMISSION = process.env.PERMISSION ?? 'active';
-export const BROADCAST = Boolean(process.env.BROADCAST ?? 'true');
+import { PUBLIC_ACTOR, PUBLIC_PERMISSION, PUBLIC_BROADCAST } from "$env/static/public";
+
+export const ACTOR = PUBLIC_ACTOR;
+export const ACCOUNT = 'eosio.faucet';
+export const FAUCET = ACCOUNT;
+export const PERMISSION = PUBLIC_PERMISSION ?? 'active';
+export const BROADCAST = Boolean(PUBLIC_BROADCAST ?? 'true');
 
 export const CHAINS = {
     jungle4: {
