+++ src/routes/api/ratelimit/+server.ts
@@ -1,7 +1,3 @@
-export const runtime = "edge";
-
-export const revalidate = 86400;
-
 export async function GET(request: Request) {
     return new Response("rate limit exceeded", {status: 429});
 }
